+++ WinQuake/view.h
@@ -30,6 +30,7 @@ extern cvar_t lcd_x;
 
 void V_Init (void);
 void V_RenderView (void);
+void V_DrawCrosshair (void);
 float V_CalcRoll (vec3_t angles, vec3_t velocity);
 void V_UpdatePalette (void);